chore: 컴포넌트 및 타입 파일로 분리
