+++ src/app/layout.tsx
@@ -17,9 +17,9 @@ export default function RootLayout({
 }>) {
   return (
     <html lang="ko">
-      <body className="min-h-screen bg-white">
+      <body className="flex min-h-screen flex-col bg-white">
         <Header />
-        <main className="mx-auto min-h-[calc(100vh-148px)] max-w-7xl px-4 py-8 sm:px-6 lg:px-8">
+        <main className="max-w-7xlitems-center flex flex-1 justify-center px-4 py-8 sm:px-6 lg:px-8">
           {children}
         </main>
         <Footer />
