+++ src/features/auth/components/SigninForm.tsx
@@ -1,0 +1,116 @@
+'use client'
+import { useState } from 'react'
+import { LoginFormData } from '@/types/auth'
+
+export default function SigninForm() {
+  // 폼 상태 관리
+  const [formData, setFormData] = useState<LoginFormData>({
+    email: '',
+    password: '',
+  })
+
+  // 비밀번호 표시 상태
+  const [showPassword, setShowPassword] = useState(false)
+
+  // 입력값 변경 핸들러
+  const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const { name, value } = e.target
+    setFormData((prev) => ({ ...prev, [name]: value }))
+  }
+
+  // 비밀번호 표시 토글
+  const togglePasswordVisibility = () => {
+    setShowPassword((prev) => !prev)
+  }
+
+  // 폼 제출 핸들러
+  const handleSubmit = (e: React.FormEvent) => {
+    e.preventDefault()
+    console.log('Login submitted:', formData)
+    // 실제 구현 시 로그인 API 호출
+  }
+  return (
+    <form onSubmit={handleSubmit} className="space-y-6">
+      {/* 이메일 입력 */}
+      <div>
+        <label htmlFor="email" className="mb-2 block text-sm font-medium text-gray-700">
+          이메일 <span className="text-red-500">*</span>
+        </label>
+        <input
+          type="email"
+          id="email"
+          name="email"
+          value={formData.email}
+          onChange={handleInputChange}
+          placeholder="example@welearn.com"
+          required
+          autoComplete="email"
+          className="w-full rounded-lg border border-gray-300 px-4 py-3 text-sm placeholder:text-gray-400 focus:border-transparent focus:ring-2 focus:ring-gray-500 focus:outline-none"
+        />
+      </div>
+
+      {/* 비밀번호 입력 */}
+      <div>
+        <label htmlFor="password" className="mb-2 block text-sm font-medium text-gray-700">
+          비밀번호 <span className="text-red-500">*</span>
+        </label>
+        <div className="relative">
+          <input
+            type={showPassword ? 'text' : 'password'}
+            id="password"
+            name="password"
+            value={formData.password}
+            onChange={handleInputChange}
+            placeholder="8자 이상 입력하세요"
+            required
+            autoComplete="current-password"
+            className="w-full rounded-lg border border-gray-300 px-4 py-3 pr-12 text-sm placeholder:text-gray-400 focus:border-transparent focus:ring-2 focus:ring-gray-500 focus:outline-none"
+          />
+          {/* 비밀번호 표시/숨김 토글 버튼 */}
+          <button
+            type="button"
+            onClick={togglePasswordVisibility}
+            className="absolute inset-y-0 right-0 flex items-center px-3 text-gray-400 transition-colors hover:text-gray-600"
+            aria-label={showPassword ? '비밀번호 숨기기' : '비밀번호 표시'}
+          >
+            {showPassword ? (
+              // 눈 감은 아이콘 (숨기기)
+              <svg className="h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="M13.875 18.825A10.05 10.05 0 0112 19c-4.478 0-8.268-2.943-9.543-7a9.97 9.97 0 011.563-3.029m5.858.908a3 3 0 114.243 4.243M9.878 9.878l4.242 4.242M9.88 9.88l-3.29-3.29m7.532 7.532l3.29 3.29M3 3l3.59 3.59m0 0A9.953 9.953 0 0112 5c4.478 0 8.268 2.943 9.543 7a10.025 10.025 0 01-4.132 5.411m0 0L21 21"
+                />
+              </svg>
+            ) : (
+              // 눈 뜬 아이콘 (표시)
+              <svg className="h-5 w-5" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="M15 12a3 3 0 11-6 0 3 3 0 016 0z"
+                />
+                <path
+                  strokeLinecap="round"
+                  strokeLinejoin="round"
+                  strokeWidth="2"
+                  d="M2.458 12C3.732 7.943 7.523 5 12 5c4.478 0 8.268 2.943 9.542 7-1.274 4.057-5.064 7-9.542 7-4.477 0-8.268-2.943-9.542-7z"
+                />
+              </svg>
+            )}
+          </button>
+        </div>
+      </div>
+
+      {/* 로그인 버튼 */}
+      <button
+        type="submit"
+        className="w-full rounded-lg bg-gray-900 px-6 py-3 text-base font-medium text-white transition-colors hover:bg-gray-800 focus:ring-2 focus:ring-gray-500 focus:ring-offset-2 focus:outline-none"
+      >
+        로그인
+      </button>
+    </form>
+  )
+}
