+++ src/types/auth.ts
@@ -1,0 +1,11 @@
+export interface LoginFormData {
+  email: string
+  password: string
+}
+
+export interface SignupFormData {
+  name: string
+  email: string
+  password: string
+  passwordConfirm: string
+}
